requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==3.2.15
django-rest-framework==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.3
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uWSGI==2.0.20
whitenoise==6.2.0
